+++ ProjetGl/ObjParser.cpp
@@ -143,8 +143,8 @@ bool loadObjFile(const char* file_path, std::vector<glm::vec4> &geometric_vertex
 
 					if (nbVertices == 3) {
 						// The face has 4 vertices, therefore we need to split it into 2 triangles
-						int index1 = m_face.vertex.size() - 4;
-						int index2 = m_face.vertex.size() - 2;
+						int index1 = (int)m_face.vertex.size() - 4;
+						int index2 = (int)m_face.vertex.size() - 2;
 						m_face.vertex.push_back(m_face.vertex.at(index1));
 						m_face.text_coords.push_back(m_face.text_coords.at(index1));
 						m_face.normals.push_back(m_face.normals.at(index1));
